aegis-system/requirements.txt
# Core Frameworks
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# AI & Agents
langgraph>=0.2.50
langchain>=0.3.10
langchain-google-genai>=2.0.5

# MCP Tools
fastmcp>=2.10.0
mcp>=1.1.0

# Database
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
pgvector>=0.3.4
pg8000
geoalchemy2>=0.15.0
alembic>=1.13.0
shapely>=2.0.0
cloud-sql-python-connector[pg8000]>=1.9.0
feedparser
networkx

# Utilities
httpx>=0.27.0
